api/requirements.txt
# promptlock API server dependencies
# Install with: pip install -r api/requirements.txt
# Or as a package extra: pip install -e ".[api]"

fastapi>=0.111
uvicorn[standard]>=0.30
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
pydantic[email]>=2.7
alembic>=1.13
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
boto3>=1.34
python-multipart>=0.0.9
llm-toolkit-schema>=1.1.1
